% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{build_cache_path_directory}
\alias{build_cache_path_directory}
\title{Build cache file path for directory data}
\usage{
build_cache_path_directory(cache_type)
}
\arguments{
\item{cache_type}{Type of cache ("directory_tidy" or "directory_raw")}
}
\value{
File path string
}
\description{
Build cache file path for directory data
}
\keyword{internal}
