% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_empty_district_df}
\alias{create_empty_district_df}
\title{Create empty district data frame}
\usage{
create_empty_district_df(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Empty data frame with district structure
}
\description{
Create empty district data frame
}
\keyword{internal}
