% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{create_empty_enrollment_df}
\alias{create_empty_enrollment_df}
\title{Create empty enrollment data frame}
\usage{
create_empty_enrollment_df()
}
\value{
Empty data frame with enrollment columns
}
\description{
Creates an empty data frame with the expected enrollment schema.
}
\keyword{internal}
