% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_pdf_era}
\alias{download_pdf_era}
\title{Download enrollment data from PDF files (2000-2007)}
\usage{
download_pdf_era(end_year)
}
\arguments{
\item{end_year}{School year end (2000-2007)}
}
\value{
Data frame with enrollment data
}
\description{
Wyoming's historical enrollment data is available as PDF files.
This function downloads and parses these PDFs.
}
\keyword{internal}
