% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Wyoming enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available years
}
\description{
Returns a vector of years for which enrollment data is available.
}
\examples{
get_available_years()
}
