% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_path}
\alias{get_cache_path}
\title{Get cache file path for given year and type}
\usage{
get_cache_path(end_year, type)
}
\arguments{
\item{end_year}{School year end}

\item{type}{Data type ("tidy" or "wide")}
}
\value{
Full path to cache file
}
\description{
Get cache file path for given year and type
}
\keyword{internal}
