% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from WDE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
Data frame with enrollment data
}
\description{
Downloads school enrollment data from Wyoming Department of Education.
Uses PDF files for 2000-2007 and the modern reporting system for 2008+.
}
\keyword{internal}
