% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_pdf_line}
\alias{parse_pdf_line}
\title{Parse a single line from enrollment PDF}
\usage{
parse_pdf_line(line, end_year)
}
\arguments{
\item{line}{Text line from PDF}

\item{end_year}{School year for context}
}
\value{
Data frame row or NULL if line is not data
}
\description{
Parse a single line from enrollment PDF
}
\keyword{internal}
