% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_wde_html_report}
\alias{parse_wde_html_report}
\title{Parse WDE HTML report}
\usage{
parse_wde_html_report(html_path, end_year)
}
\arguments{
\item{html_path}{Path to the HTML file}

\item{end_year}{School year for context}
}
\value{
Data frame with enrollment data
}
\description{
Extracts tabular data from WDE's HTML report output.
}
\keyword{internal}
