% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{read_cache_directory}
\alias{read_cache_directory}
\title{Read directory data from cache}
\usage{
read_cache_directory(cache_type)
}
\arguments{
\item{cache_type}{Type of cache ("directory_tidy" or "directory_raw")}
}
\value{
Cached data frame
}
\description{
Read directory data from cache
}
\keyword{internal}
