% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_cache}
\alias{write_cache}
\title{Write data to cache}
\usage{
write_cache(df, end_year, type)
}
\arguments{
\item{df}{Data frame to cache}

\item{end_year}{School year end}

\item{type}{Data type ("tidy" or "wide")}
}
\value{
Invisibly returns the cache path
}
\description{
Write data to cache
}
\keyword{internal}
