% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{write_cache_directory}
\alias{write_cache_directory}
\title{Write directory data to cache}
\usage{
write_cache_directory(data, cache_type)
}
\arguments{
\item{data}{Data frame to cache}

\item{cache_type}{Type of cache ("directory_tidy" or "directory_raw")}
}
\value{
Invisibly returns the cache path
}
\description{
Write directory data to cache
}
\keyword{internal}
